+++ src/pat/masonry/masonry.js
@@ -43,7 +43,11 @@
         init: function masonryInit($el, opts) {
             this.options = parser.parse(this.$el, opts);
             this.initMasonry();
-            imagesLoaded(this.$el, this.layout.bind(this));
+            var imgLoad = imagesLoaded(this.$el);
+            imgLoad.on("progress", function() {
+                this.msnry.layout();
+            }.bind(this));
+            imgLoad.on("done", this.layout.bind(this));
             // Update if something gets injected inside the pat-masonry
             // element.
             this.$el.on("patterns-injected.pat-masonry",